Migrate topology service to typescript.
